requirements.txt
Flask==3.1.2
deepface==0.0.95
opencv-python==4.12.0.88
tensorflow==2.20.0
tf-keras==2.20.1
Pillow==12.0.0
numpy==2.2.6
Werkzeug==3.1.3